viikon tehtävät tehty

+++ controllers/blogs.js
@@ -1,8 +1,13 @@
 const blogRouter=require('express').Router()
 const Blog=require('../models/blog')
+const User=require('../models/user')
+const jwt=require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response) => {
-  const blogs= await Blog.find({})
+  const blogs= await Blog
+    .find({})
+    .populate('user', {username: 1, name:1} )
   response.json(blogs)
 })
 
@@ -21,21 +26,36 @@ blogRouter.get('/:id', async (request, response)=>{
 })
 
 blogRouter.post('/', async (request, response) => {
+  const body = request.body
   try{
-    const body = request.body
+    const decodedToken=jwt.verify(request.token, process.env.SECRET)
+    console.log(request.token)
+
+    if(!request.token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
 
     if(body.title===undefined || body.url === undefined) {
       return response.status(400).json({error: 'bad request'})
     }
 
+    const user=await User.findById(decodedToken.id)
+
     const blog=new Blog({
       title: body.title,
-      author: body.author,
+      author: user.name,
       url: body.url,
-      likes: body.likes===undefined ? 0 : body.likes
+      likes: body.likes===undefined ? 0 : body.likes,
+      user:user._id
     })
 
     const savedBlog=await blog.save()
+    console.log(user._id)
+    console.log(savedBlog._id)
+
+    user.blogs=user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
   } catch(er) {
     console.log(er)
@@ -44,6 +64,17 @@ blogRouter.post('/', async (request, response) => {
 })
 blogRouter.delete('/:id', async (request, response)=>{
   try{
+    const decodedToken=jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+    const userSigned=await User.findById(decodedToken.id)
+    const blog=await Blog.findById(request.params.id)
+    if(!(blog.user.toString()===userSigned._id.toString())){
+      return response.status(400).json({error: 'only author can remove a blog'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
